remove hw/i386, hw/mips.
