--- qemu/hw/i386/Makefile.objs
@@ -1,1 +1,0 @@
-obj-y += pc.o
