+++ qemu/hw/ppc/ppc_booke.c
@@ -21,14 +21,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-#include "hw/hw.h"
 #include "hw/ppc/ppc.h"
 #include "qemu/timer.h"
 #include "sysemu/sysemu.h"
-//#include "hw/timer/m48t59.h"
 #include "qemu/log.h"
-//#include "hw/loader.h"
-//#include "kvm_ppc.h"
 
 
 /* Timer Control Register */
