+++ qemu/target-arm/Makefile.objs
@@ -1,6 +1,6 @@
 obj-y += translate.o op_helper.o helper.o cpu.o
 obj-y += neon_helper.o iwmmxt_helper.o
-obj-$(CONFIG_SOFTMMU) += psci.o
+obj-y += psci.o
 obj-$(TARGET_AARCH64) += cpu64.o translate-a64.o helper-a64.o unicorn_aarch64.o
 obj-$(TARGET_ARM) += unicorn_arm.o
 obj-y += crypto_helper.o
