+++ qemu/target-i386/smm_helper.c
@@ -52,7 +52,10 @@ void do_smm_enter(X86CPU *cpu)
     log_cpu_state_mask(CPU_LOG_INT, CPU(cpu), CPU_DUMP_CCOP);
 
     env->hflags |= HF_SMM_MASK;
+#if 0
+    /* not implemented. cpu_smm_update(env) does nothing. */
     cpu_smm_update(env);
+#endif
 
     sm_state = env->smbase + 0x8000;
 
@@ -308,7 +311,10 @@ void helper_rsm(CPUX86State *env)
     }
 #endif
     env->hflags &= ~HF_SMM_MASK;
+#if 0
+    /* not implemented. cpu_smm_update(env) does nothing. */
     cpu_smm_update(env);
+#endif
 
     qemu_log_mask(CPU_LOG_INT, "SMM: after RSM\n");
     log_cpu_state_mask(CPU_LOG_INT, CPU(cpu), CPU_DUMP_CCOP);
